+++ grammars/fallout-ssl/queries/highlights.scm
@@ -115,6 +115,7 @@
 ; ----- Literals -----
 
 (string) @string
+(char) @character
 (number) @number
 (boolean) @constant.builtin